requirements.txt
altair==5.4.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
appnope==0.1.3
arrow==1.3.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
black==22.12.0
blinker==1.9.0
cachetools==5.5.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
coverage==7.4.4
cryptography==42.0.2
cycler==0.12.1
dateutils==0.6.12
debugpy==1.8.0
decorator==5.1.1
dill==0.3.9
distro==1.9.0
dj-database-url==1.3.0
Django==4.2.11
django-widget-tweaks==1.5.0
docutils==0.21.2
duckdb==1.1.2
et-xmlfile==1.1.0
executing==2.0.1
Flask==3.1.0
fonttools==4.49.0
forge==0.22.0
forge-core==0.4.0
forge-db==0.3.4
forge-heroku==0.4.0
forge-tailwind==0.3.4
forge-work==0.6.0
fpdf==1.7.2
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
gunicorn==20.1.0
gw_dsl_parser==0.1.49.1
h11==0.14.0
hiredis==2.3.2
honcho==1.1.0
httpcore==1.0.7
httpx==0.27.2
hupper==1.12.1
idna==3.6
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.28.0
ipylab==1.0.0
ipython==8.20.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.7.1
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-summarytools==0.2.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.13
kanaries_track==0.0.5
Kivy==2.3.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mypy-extensions==1.0.0
narwhals==1.14.1
nest-asyncio==1.5.9
numpy==1.26.3
openai==1.55.0
openpyxl==3.1.2
packaging==23.2
pandas==1.5.3
parso==0.8.3
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==5.28.3
psutil==5.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.17.2
pygwalker==0.4.9.10
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
pyparsing==3.1.2
pytest==7.4.4
pytest-django==4.8.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2024.1
pyzmq==25.1.2
quickjs==1.19.4
redis==4.6.0
referencing==0.35.1
requests==2.31.0
rich==13.9.4
rpds-py==0.21.0
ruff==0.0.194
scapy==2.5.0
scikit-learn==1.5.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
segment-analytics-python==2.2.3
setuptools==69.0.3
six==1.16.0
sklearn2pmml==0.111.1
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlglot==25.26.0
sqlparse==0.4.4
stack-data==0.6.3
statsmodels==0.14.1
streamlit==1.40.1
tenacity==9.0.0
threadpoolctl==3.3.0
toml==0.10.2
tornado==6.4
tqdm==4.67.0
traitlets==5.14.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.1
wasmtime==25.0.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.6.0
widgetsnbextension==4.0.13
xgboost==2.1.2
